profile 0.8

+++ src/validator/ServiceUser/serviceSoldCreatebasic.js
@@ -8,14 +8,15 @@ const ValidateResult = (req, res, next) => {
         validationResult(req).throw()
         return next()
     } catch (err) {
-        fs.unlink(req.file.path, (err) => {
-            if (err) {
-                return res.status(500).json({
-                    res: "error server"
-                })
-            }
-            console.log('Archivo eliminado con éxito');
-        });
+        if (req.file) {
+            fs.unlink(req.file.path, (err) => {
+                if (err) {
+                    return res.status(500).json({
+                        res: "error server"
+                    })
+                }
+            });
+        }
         return res.status(403).json({ res: err.array() })
     }
 }
